+++ src/components/manage/Blocks/Search/components/index.js
@@ -5,3 +5,6 @@ export CheckboxFacet from './CheckboxFacet';
 export SearchInput from './SearchInput';
 export FilterList from './FilterList';
 export SortOn from './SortOn';
+export ToggleFacet from './ToggleFacet';
+export SelectFacetFilterListEntry from './SelectFacetFilterListEntry';
+export ToggleFacetFilterListEntry from './ToggleFacetFilterListEntry';