+++ src/helpers/Extensions/index.js
@@ -1,2 +1,2 @@
 export * from './withBlockSchemaEnhancer';
-export withBlockExtensions from './withBlockExtensions';
+export withBlockExtensions, { resolveExtension } from './withBlockExtensions';